DEV: Move the drag test reset out of the production bundle
`resetDragAndDropForTesting` only dispatches a `dragend`, so unlike the sibling
reset helpers it holds no module-private state and does not have to sit beside
the code it resets. It moves to the drag-and-drop test helper and its module
leaves the application bundle.
